introduce large instance classes
