+++ .circleci/config.yml
@@ -36,9 +36,16 @@ jobs:
         description: |
           The test suite requires that Azure AKS cluster to be available.
         type: boolean
+      resource-class:
+        default: 'medium'
+        description: |
+          CircleCI instance size
+        type: string
 
     machine:
       image: ubuntu-1604:202007-01
+
+    resource_class: resource-class
 
     environment:
       HELM_VERSION: "v3.6.3"
@@ -137,7 +144,12 @@ workflows:
           tag: "upgrade"
           requires-minikube: true
       - go-test:
+          name: "Large Instance Tests"
+          test-suite: "./test/minikube"
+          tag: "large"
+          requires-minikube: true
+          resource-class: large
+      - go-test:
           name: "Multi-cluster tests"
           test-suite: "./test/multicluster"
           requires-aks: true
-
